+++ src/main/java/net/vexelon/currencybg/srv/api/Currencies.java
@@ -16,9 +16,9 @@ import org.slf4j.LoggerFactory;
 import com.google.common.collect.Maps;
 
 import net.vexelon.currencybg.srv.Defs;
-import net.vexelon.currencybg.srv.db.DataSource;
+import net.vexelon.currencybg.srv.db.MySQLDataSource;
 import net.vexelon.currencybg.srv.db.DataSourceException;
-import net.vexelon.currencybg.srv.db.DataSourceInterface;
+import net.vexelon.currencybg.srv.db.DataSource;
 import net.vexelon.currencybg.srv.utils.DateTimeUtils;
 import net.vexelon.currencybg.srv.utils.DateTimeUtilsNew;
 
@@ -46,9 +46,9 @@ public class Currencies extends AbstractJunction {
 	@GET
 	@Path("/{dateFrom}")
 	public Response getAllRatesByDate(@PathParam("dateFrom") String dateFrom,
-			@HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
+	        @HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
 
-		try (DataSourceInterface source = new DataSource()) {
+		try (DataSource source = new MySQLDataSource()) {
 			verifyAccess();
 
 			source.connect();
@@ -69,9 +69,9 @@ public class Currencies extends AbstractJunction {
 	@GET
 	@Path("/{dateFrom}/{sourceId}")
 	public Response getAllRates(@PathParam("dateFrom") String dateFrom, @PathParam("sourceId") Integer sourceId,
-			@HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
+	        @HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
 
-		try (DataSourceInterface source = new DataSource()) {
+		try (DataSource source = new MySQLDataSource()) {
 			verifyAccess();
 
 			source.connect();
@@ -93,9 +93,9 @@ public class Currencies extends AbstractJunction {
 	@GET
 	@Path("/today/{timeFrom}/{sourceId}")
 	public Response getAllCurrentRatesAfter(@PathParam("timeFrom") String timeFrom,
-			@PathParam("sourceId") Integer sourceId, @HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
+	        @PathParam("sourceId") Integer sourceId, @HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
 
-		try (DataSourceInterface source = new DataSource()) {
+		try (DataSource source = new MySQLDataSource()) {
 			verifyAccess();
 
 			source.connect();
@@ -113,7 +113,7 @@ public class Currencies extends AbstractJunction {
 			String localTimeFromNoTz = DateTimeUtilsNew.removeTimeZone(timeFrom, Defs.DATETIME_FORMAT);
 
 			return getJsonResponse(source.getAllCurrentRatesAfter(sourceId,
-					DateTimeUtils.parseDate(localTimeFromNoTz, Defs.DATETIME_FORMAT)));
+			        DateTimeUtils.parseDate(localTimeFromNoTz, Defs.DATETIME_FORMAT)));
 		} catch (IOException | DataSourceException | ParseException e) {
 			log.error("", e);
 			return getErrorResponse();
@@ -126,9 +126,9 @@ public class Currencies extends AbstractJunction {
 	@GET
 	@Path("/today/{timeFrom}")
 	public Response getAllCurrentRatesAfter(@PathParam("timeFrom") String timeFrom,
-			@HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
+	        @HeaderParam(Defs.HEADER_APIKEY) String apiKey) throws Exception {
 
-		try (DataSourceInterface source = new DataSource()) {
+		try (DataSource source = new MySQLDataSource()) {
 			verifyAccess();
 
 			source.connect();
@@ -145,7 +145,7 @@ public class Currencies extends AbstractJunction {
 			String localTimeFromNoTz = DateTimeUtilsNew.removeTimeZone(timeFrom, Defs.DATETIME_FORMAT);
 
 			return getJsonResponse(
-					source.getAllCurrentRatesAfter(DateTimeUtils.parseDate(localTimeFromNoTz, Defs.DATETIME_FORMAT)));
+			        source.getAllCurrentRatesAfter(DateTimeUtils.parseDate(localTimeFromNoTz, Defs.DATETIME_FORMAT)));
 		} catch (IOException | DataSourceException | ParseException e) {
 			log.error("", e);
 			return getErrorResponse();