+++ HISTORY.md
@@ -1,6 +1,14 @@
 Currency BG Server Changes
 ==========================
 
+## 1.1
+
+ * Added sending error reports to configurable E-mail addresses via SparkPost (Ref #48)
+ * Added factorin.bg currencies source (Ref #44)
+ * Fixed Let's Encrypt SSL errors for HTTPS domains (Ref #56)
+ * Fixed "I/O dispatcher" httpasyncclient errors (Ref #59)
+ * Fixed Tomcat JDBC memory leak errors (Ref #28)
+  
 ## 1.0
 
-Initial release.
+ * Initial release.